fixed insertBlobber handling for conductor
